Add support for WordPress Multisite installations.
The current version of the plugin appears in the Tools menu of every
site within a WordPress Multisite installation. In the case of a
Multisite installation, the administration options should only be
available to users with the Manage Network Options capability, and the
settings page should show within the Network section of the admin panel.