Check certain Core functions exist before using them
If used in conjunction with a page cache like Batcache, the plugin currently triggers fatal errors because `advanced-cache.php` is loaded very early, before most of Core is available. Interacting with the object cache at this early stage is supported by Core, but does mean that not all Core functions can be used in the object class itself.