WP_REDIS_DISABLE_DROPIN_MANAGE
I had previously done a PR to hide the dropin banner. This code was not complete. I need a wider deactivation of the dropin management. In my case, I want to master the dropin code to install and when. So I replaced my constant by WP_REDIS_DISABLE_DROPIN_MANAGE