Update object-cache.php
The above code ensure a unique key salt prefix for customers hosting multiple sites on the same Cloudways server, the code checks if the site is hosted on a Cloudways environment along with if the WP_CACHE_KEY_SALT is defined in wp-config.php or not.  if key prefix is not defined and site is hosted on Cloudways the unique key-prefix will be set automatically by the plugin.