fix for when cache item's value is set to null (#381)
If a cache item has a value of "null", the current isset() mechanism returns false. This results in the cache key never being used and, in my case, hundreds of duplicate queries on each page load. This changes it to array_key_exists. Likewise, it would never delete the cache key, and probably other issues.